% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_list_to_body.R
\name{add_list_to_body}
\alias{add_list_to_body}
\title{Add a formatted list of items to a \code{rdocx} object}
\usage{
add_list_to_body(
  docx,
  values,
  style = "List Bullet",
  keep_na = FALSE,
  before = NULL,
  after = "",
  ...
)
}
\arguments{
\item{docx}{A \code{rdocx} object.}

\item{values}{Character vector with values for list.}

\item{style}{Style to use for list of values, Default: 'List Bullet'}

\item{keep_na}{If \code{TRUE}, keep \code{NA} values in values Default: FALSE}

\item{before}{Value to insert before the list, Default: NULL}

\item{after}{Value to insert after the list, Default: ''}

\item{...}{Additional parameters passed to \code{\link[=cursor_docx]{cursor_docx()}}}
}
\value{
A modified \code{rdocx} object.
}
\description{
This function assumes that the input docx object has a list style. Use
\code{\link[=read_docx_ext]{read_docx_ext()}} with \code{allow_null = TRUE} to create a new rdocx object with
the "List Bullet" and "List Number" styles. "List Paragraph" can be used but
the resulting formatted list will not have bullets or numbers.
}
\seealso{
\code{\link[=add_to_body]{add_to_body()}}, \code{\link[=vec_add_to_body]{vec_add_to_body()}}
}
