% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{check_docx}
\alias{check_docx}
\alias{check_docx_fileext}
\title{Check if x is a rdocx object or if x has a docx file extension}
\usage{
check_docx(x, arg = caller_arg(x), ...)

check_docx_fileext(x, arg = caller_arg(x), ..., .envir = current_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name for x. Defaults to \code{caller_arg(x)}.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Check if x is a rdocx object or if x has a docx file extension
}
