% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{check_office_fileext}
\alias{check_office_fileext}
\alias{check_docx_fileext}
\alias{check_pptx_fileext}
\alias{check_xlsx_fileext}
\title{Check file path for a docx, pptx, or xlsx file extension}
\usage{
check_office_fileext(
  x,
  arg = caller_arg(x),
  fileext = c("docx", "pptx", "xlsx"),
  allow_null = FALSE,
  call = caller_env(),
  ...
)

check_docx_fileext(x, arg = caller_arg(x), call = caller_env(), ...)

check_pptx_fileext(x, arg = caller_arg(x), call = caller_env(), ...)

check_xlsx_fileext(x, arg = caller_arg(x), call = caller_env(), ...)
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name of object to check. Used to improve
\code{\link[cli:cli_abort]{cli::cli_abort()}} messages. Defaults to \code{caller_arg(x)}.}

\item{fileext}{File extensions to allow without error. Defaults to "docx",
"pptx", "xlsx".}

\item{allow_null}{If \code{TRUE} and x is \code{NULL}, invisibly return \code{NULL}. If
\code{FALSE} (default), error unless x is a character vector with file extension
matching the supplied fileext value.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\value{
Invisibly returns \code{NULL} if \code{x} passes the check (or is \code{NULL}
and \code{allow_null} is \code{TRUE}). Otherwise, errors.
}
\description{
Check file path for a docx, pptx, or xlsx file extension
}
