% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_docx.R
\name{convert_docx}
\alias{convert_docx}
\title{Convert an rodcx object or Word file to another format with pandoc}
\usage{
convert_docx(
  docx = NULL,
  to = NULL,
  input = NULL,
  output = NULL,
  path = NULL,
  options = NULL,
  extract_media = TRUE,
  overwrite = TRUE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{docx}{A rdocx object or path to a docx file. Optional if input is
provided.}

\item{to}{Format to convert to (if not specified, you must specify
\code{output})}

\item{input}{Character vector containing paths to input files
(files must be UTF-8 encoded)}

\item{output}{Output file (if not specified then determined based on format
being converted to).}

\item{path}{File path for converted file passed to wd parameter of
\code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}. If docx is a rdocx object, path defaults to
current working directory instead of the base directory of the input file
path.}

\item{options}{Character vector of command line options to pass to pandoc.}

\item{extract_media}{If \code{TRUE}, append \code{"--extract-media=."} to
options. Defaults to \code{FALSE}.}

\item{overwrite}{If \code{TRUE} (default), remove file at path if it already
exists. If \code{FALSE} and file exists, this function aborts.}

\item{quiet}{If \code{TRUE}, suppress alert messages and pass \code{FALSE} to verbose
parameter of \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}. Defaults \code{TRUE}.}

\item{...}{
  Arguments passed on to \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert}}
  \describe{
    \item{\code{from}}{Format to convert from (if not specified then the format is
determined based on the file extension of \code{input}).}
    \item{\code{citeproc}}{\code{TRUE} to run the pandoc-citeproc filter (for processing
citations) as part of the conversion.}
  }}
}
\value{
Executes a call to pandoc using \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}} to
create a file from an officer object or a docx file.
}
\description{
\code{convert_docx()} uses \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}} to convert a rodcx object
or Word file to another format with pandoc. If the "to" parameter contains a
file extension, it is assumed to be an output file name. If you want to
convert a file to a Word document, use the input parameter for the path to
the Markdown, HTML, or other file.
}
\details{
Supported input and output formats are described in the
\href{https://pandoc.org/MANUAL.html}{pandoc user guide}.

The system path as well as the version of pandoc shipped with RStudio (if
running under RStudio) are scanned for pandoc and the highest version
available is used.
}
\examples{
docx_example <- read_officer(
  system.file("doc_examples/example.docx", package = "officer")
)

convert_docx(
  docx_example,
  to = "markdown"
)

withr::with_tempdir({
  convert_docx(
    docx_example,
    output = "docx_example.html"
  )
})
}
\seealso{
\code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}
}
