% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_convert.R
\name{docx_convert}
\alias{docx_convert}
\title{Convert an rodcx object or Word file to another format with pandoc}
\usage{
docx_convert(
  docx,
  to = NULL,
  input = NULL,
  output = NULL,
  path = NULL,
  quiet = TRUE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{docx}{A rdocx object or path to a docx file. Optional if input is
provided.}

\item{to}{Format to convert to (if not specified, you must specify
\code{output})}

\item{input}{Character vector containing paths to input files
(files must be UTF-8 encoded)}

\item{output}{Output file (if not specified then determined based on format
being converted to).}

\item{path}{File path.}

\item{quiet}{If \code{TRUE}, suppress alert messages and pass \code{FALSE} to verbose
parameter of \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}. Defaults \code{TRUE}.}

\item{overwrite}{If \code{TRUE} (default), remove file at path if it already
exists. If \code{FALSE} and file exists, this function aborts.}

\item{...}{
  Arguments passed on to \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert}}
  \describe{
    \item{\code{from}}{Format to convert from (if not specified then the format is
determined based on the file extension of \code{input}).}
    \item{\code{citeproc}}{\code{TRUE} to run the pandoc-citeproc filter (for processing
citations) as part of the conversion.}
    \item{\code{options}}{Character vector of command line options to pass to pandoc.}
    \item{\code{wd}}{Working directory in which code will be executed. If not
supplied, defaults to the common base directory of \code{input}.}
  }}
}
\value{
Executes a call to pandoc using \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}} to
create a file from an officer object or a docx file.
}
\description{
Use \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}} to convert a rodcx object or Word file to
another format with pandoc.
}
\details{
Supported input and output formats are described in the
\href{https://pandoc.org/MANUAL.html}{pandoc user guide}.

The system path as well as the version of pandoc shipped with RStudio (if
running under RStudio) are scanned for pandoc and the highest version
available is used.
}
\examples{
\dontrun{
if (interactive()) {
  docx <- read_officer(
    system.file("doc_examples/example.docx", package = "officer")
  )

  withr::with_tempdir({
    docx_convert(
      docx,
      output = "docx.md"
    )
  })
}
}
}
\seealso{
\code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}}
}
