% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{has_docx_fileext}
\alias{has_docx_fileext}
\title{Does x have a docx file extension?}
\usage{
has_docx_fileext(string = NULL, ignore.case = TRUE)
}
\arguments{
\item{string}{String to be tested with or without filetype. Defaults to
\code{NULL}.}

\item{ignore.case}{If \code{FALSE}, the pattern matching is case sensitive. If
\code{TRUE}, case is ignored.}
}
\description{
Does x have a docx file extension?
}
\keyword{internal}
