% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{list_doc_properties}
\alias{list_doc_properties}
\title{List document properties for a rdocx or rpptx object}
\usage{
list_doc_properties(docx, filename = NULL)
}
\description{
List document properties for a rdocx or rpptx object
}
\keyword{internal}
