% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_list.R
\name{make_block_list}
\alias{make_block_list}
\alias{combine_blocks}
\title{Make a list of blocks to add to a Word document or PowerPoint presentation}
\usage{
make_block_list(blocks = NULL, ..., allow_empty = FALSE, call = caller_env())

combine_blocks(...)
}
\arguments{
\item{blocks}{A list of parameters to pass to \code{\link[officer:block_list]{officer::block_list()}} or a
\code{block_list} object. If parameters are provided to both \code{...} and blocks is
a \code{block_list}, the additional parameters are appended to the end of
blocks.}

\item{...}{For \code{\link[=make_block_list]{make_block_list()}}, these parameters are passed to
\code{\link[officer:block_list]{officer::block_list()}} and must \emph{not} include \code{block_list} objects. For
\code{\link[=combine_blocks]{combine_blocks()}}, these parameters must all be \code{block_list} objects.}

\item{allow_empty}{If \code{TRUE}, \code{\link[=check_block_list]{check_block_list()}} allows an empty block
list. Defaults to \code{FALSE}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{make_block_list()} extends \code{\link[officer:block_list]{officer::block_list()}} by supporting a list of
inputs and optionally combining parameters with an existing block list (using
the blocks parameter). Unlike \code{\link[officer:block_list]{officer::block_list()}}, \code{\link[=make_block_list]{make_block_list()}}
errors if no input parameters are provided.

\code{combine_blocks()} takes any number of \code{block_list} objects and combined them
into a single \code{block_list}. Both functions are not yet working as expected.
}
\seealso{
Other block list functions:
\code{\link[=officer_add_blocks]{officer_add_blocks()}}
}
\concept{block list functions}
