% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blocks.R
\name{officer_add_blocks}
\alias{officer_add_blocks}
\alias{add_blocks_to_body}
\title{Add a list of blocks into a Word document or PowerPoint presentation}
\usage{
officer_add_blocks(
  x,
  blocks,
  pos = "after",
  location = NULL,
  level_list = integer(0),
  ...,
  call = caller_env()
)

add_blocks_to_body(
  docx,
  blocks,
  pos = "after",
  keyword = NULL,
  id = NULL,
  index = NULL,
  ...
)
}
\arguments{
\item{x}{A rdocx or rpptx object. Required.}

\item{blocks}{set of blocks to be used as footnote content returned by
function \code{\link[officer:block_list]{block_list()}}.}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{location}{If \code{NULL}, location defaults to \code{\link[officer:ph_location_type]{officer::ph_location_type()}}}

\item{level_list}{The list of levels for hierarchy structure as integer values.
If used the object is formated as an unordered list. If 1 and 2,
item 1 level will be 1, item 2 level will be 2.}

\item{...}{further arguments passed to or from other methods. When
adding a \code{ggplot} object or \code{plot_instr}, these arguments will be used
by the png function.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{docx}{A rdocx object.}

\item{keyword, id}{A keyword string used to place cursor with
\code{\link[officer:cursor_reach]{officer::cursor_reach()}} or bookmark id with \code{\link[officer:cursor_bookmark]{officer::cursor_bookmark()}}.
Defaults to \code{NULL}. If keyword or id are not provided, the gt object is
inserted at the front of the document.}

\item{index}{A integer matching a doc_index value appearing in a summary of
the docx object created with \code{\link[officer:docx_summary]{officer::docx_summary()}}. If index is for a
paragraph value, the text of the pargraph is used as a keyword.}
}
\value{
A \code{rdocx} or \code{rpptx} object with the block list added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=officer_add_blocks]{officer_add_blocks()}} supports adding a list of blocks to rdocx (using
\code{\link[=add_blocks_to_body]{add_blocks_to_body()}}) or rpptx objects (using \code{\link[officer:ph_with]{officer::ph_with()}}).
\code{add_blocks_to_body()} is a variant of \code{\link[officer:body_add_blocks]{officer::body_add_blocks()}} that
allows users to set the cursor position before adding the block list using
\code{\link[=cursor_docx]{cursor_docx()}}.
}
\seealso{
Other block list functions:
\code{\link[=make_block_list]{make_block_list()}}
}
\concept{block list functions}
