% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_media.R
\name{officer_media}
\alias{officer_media}
\title{Copy media from a docx or pptx file or a rdocx or rpptx object to a target
folder}
\usage{
officer_media(
  filename = NULL,
  path = NULL,
  x = NULL,
  target = "media",
  list = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{filename, path}{File name and path for a \code{docx} or \code{pptx} file. One of
the two must be provided. Ignored if x is provided. Defaults to \code{NULL}.}

\item{x}{A rdocx or rpptx object containing one or more media file.}

\item{target}{Folder name or path to copy media files. dir is created if no
folder exists at that location.}

\item{list}{If \code{TRUE}, display a message listing files contained in the docx
or pptx file but do not copy the files to dir. Defaults to \code{FALSE}.}

\item{overwrite}{If \code{TRUE} (default), overwrite any files with the same names
at target path. If \code{FALSE}, abort if files with the same names already
exist.}
}
\value{
Invisibly returns the paths of the copied (or, if \code{list} is
\code{TRUE}, listed) media files, or \code{NULL} if no media files are found.
}
\description{
Unzip a docx or pptx file to a temporary directory, check if the directory
contains a media folder, and copy media files to the directory set by dir. If
a rdocx or rpptx object is provided, files are copied from the temporary
package_dir associated with the object (accessible via \code{x[["package_dir"]]}).
}
\examples{
officer_media(
  system.file("doc_examples/example.pptx", package = "officer"),
  list = TRUE
)
}
\seealso{
\code{\link[officer:media_extract]{officer::media_extract()}}
}
