% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{officer_open}
\alias{officer_open}
\title{Preview a rdocx, rpptx, or rxlsx object in local default applications}
\usage{
officer_open(
  x,
  path = NULL,
  ...,
  overwrite = FALSE,
  interactive = is_interactive()
)
}
\arguments{
\item{x}{A rdocx, rpptx, or rxlsx object to save. Document properties won't
be set for rxlsx objects.}

\item{path}{Optional. Set as temporary file with file extension matching type
of input object \code{x}.}

\item{...}{
  Arguments passed on to \code{\link{write_officer}}
  \describe{
    \item{\code{modified_by}}{If the withr package is installed, modified_by overrides
the default value for the lastModifiedBy property assigned to the output
file by officer. Defaults to \code{Sys.getenv("USER")} (the same value used by
officer).}
  }}

\item{overwrite}{Defaults to \code{FALSE}.}

\item{interactive}{If \code{FALSE}, warn the user and return \code{x}.}
}
\value{
Input object \code{x} without modification.
}
\description{
\code{officer_open()} uses \code{officer::open_file()} to open a file ceated from a
rdocx, rpptx, or rxlsx object.
}
