% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_tables.R
\name{officer_tables}
\alias{officer_tables}
\alias{officer_table}
\title{Get tables from a rdocx or rpptx object}
\usage{
officer_tables(
  x,
  index = NULL,
  has_header = TRUE,
  col = NULL,
  preserve = FALSE,
  ...,
  stack = FALSE,
  type_convert = FALSE,
  nm = NULL,
  call = caller_env()
)

officer_table(
  x,
  index = NULL,
  has_header = TRUE,
  col = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{x}{A rdocx or rpptx object or a data frame created with
\code{\link[=officer_summary]{officer_summary()}}.}

\item{index}{A index value matching a doc_index value for a table in the
summary data frame, Default: \code{NULL}}

\item{has_header}{If \code{TRUE} (default), tables are expected to have implicit
headers even if the Word table does not have an explicit header row. If
\code{FALSE}, only explicit header rows will be used as column names.}

\item{col}{If col is supplied, \code{\link[=officer_table]{officer_table()}} passes col and the
additional parameters in ... to \code{\link[=fill_with_pattern]{fill_with_pattern()}}. This allows the
addition of preceding headings or captions as a column within the
data.frame returned by \code{\link[=officer_tables]{officer_tables()}}. This is an experimental feature
and may be modified or removed. Defaults to \code{NULL}.}

\item{preserve}{If \code{FALSE} (default), text in table cells is collapsed into a
single line. If \code{TRUE}, line breaks in table cells are preserved as a "\\n"
character. This feature is adapted from \code{docxtractr::docx_extract_tbl()}
published under a \href{https://github.com/hrbrmstr/docxtractr/blob/master/LICENSE}{MIT licensed} in
the 'docxtractr' package by Bob Rudis.}

\item{...}{Additional parameters passed to \code{\link[=fill_with_pattern]{fill_with_pattern()}}.}

\item{stack}{If \code{TRUE} and all tables share the same number of columns,
return a single combined data frame instead of a list. Defaults to \code{FALSE}.}

\item{type_convert}{If \code{TRUE}, convert columns for the returned data frames
to the appropriate type using \code{\link[utils:type.convert]{utils::type.convert()}}.}

\item{nm}{Names to use for returned list of tables. If \code{NULL} (default), the
names are set to the table_index values using the pattern
"table_index_<table_index_number>".}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\value{
A list of data frames or, if stack is \code{TRUE}, a single data frame.

A single data.frame, or (if the table has more than one header
row) a list of two data.frames: the table body and the table header.
}
\description{
Get one or more tables from a rdocx or rpptx object. \code{\link[=officer_tables]{officer_tables()}}
returns a list of data frames and \code{\link[=officer_table]{officer_table()}} returns a single table as
a data frame. These functions are based on example code on extracting Word
document and PowerPoint slides in the \href{https://ardata-fr.github.io/officeverse/extract-content.html#word-tables}{officeverse documentation}.
Some additional features including the type_convert parameter and the
addition of doc_index values as the default names for the returned list of
tables are based on \href{https://www.rostrum.blog/2023/06/07/rectangular-officer/}{this blog post by Matt Dray}.
}
\examples{
docx_example <- read_docx_ext(
  filename = "example.docx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(docx_example)

pptx_example <- read_pptx_ext(
  filename = "example.pptx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(pptx_example)[[1]]

}
\seealso{
\code{\link[docxtractr:docx_extract_all]{docxtractr::docx_extract_all()}}
}
