% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerExtras-package.R
\docType{package}
\name{officerExtras-package}
\alias{officerExtras}
\alias{officerExtras-package}
\title{officerExtras: Extra Helpers for 'officer'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helper and convenience functions using the 'officer' package to modify docx files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/officerExtras}
  \item \url{https://elipousson.github.io/officerExtras/}
  \item Report bugs at \url{https://github.com/elipousson/officerExtras/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
