% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{read_docx_ext}
\alias{read_docx_ext}
\title{Read a docx file or use a rdox object if provided}
\usage{
read_docx_ext(filename = NULL, path = NULL, docx = NULL, quiet = TRUE)
}
\arguments{
\item{filename, path}{File name and path. Default: \code{NULL}}

\item{docx}{rdocx class object, If docx is provided, filename and path are
ignored. Default: \code{NULL}}

\item{quiet}{If \code{FALSE}, warn if docx is provided when filename and/or path
are also provided. Default: \code{TRUE}.}
}
\value{
A rdocx object.
}
\description{
Read a docx file or use a rdox object if provided
}
\seealso{
\code{\link[officer:read_docx]{officer::read_docx()}}
}
