% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_officer.R
\name{read_officer}
\alias{read_officer}
\alias{read_docx_ext}
\alias{read_pptx_ext}
\alias{read_xlsx_ext}
\title{Read a docx, pptx, potx, or xlsx file or use an existing object from officer
if provided}
\usage{
read_officer(
  filename = NULL,
  path = NULL,
  fileext = c("docx", "pptx", "xlsx"),
  x = NULL,
  arg = caller_arg(x),
  quiet = TRUE,
  ...
)

read_docx_ext(filename = NULL, path = NULL, docx = NULL, quiet = TRUE)

read_pptx_ext(filename = NULL, path = NULL, pptx = NULL, quiet = TRUE)

read_xlsx_ext(filename = NULL, path = NULL, xlsx = NULL, quiet = TRUE)
}
\arguments{
\item{filename, path}{File name and path. Default: \code{NULL}. Must include a
"docx", "pptx", or "xlsx" file path. "dotx" and "potx" files are also
supported.}

\item{x}{A rdocx, rpptx, or rxlsx class object, If docx is provided, filename
and path are ignored. Default: \code{NULL}}

\item{quiet}{If \code{FALSE}, warn if docx is provided when filename and/or path
are also provided. Default: \code{TRUE}.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}.}

\item{docx, pptx, xlsx}{A rdocx, rpptx, or rxlsx class object passed to the x
parameter by the corresponding function. Defaults to \code{NULL}.}
}
\value{
A rdocx, rpptx, or rxlsx object.
}
\description{
\code{\link[=read_docx_ext]{read_docx_ext()}}, \code{\link[=read_pptx_ext]{read_pptx_ext()}}, and \code{\link[=read_xlsx_ext]{read_xlsx_ext()}} are wrappers for
\code{\link[=read_officer]{read_officer()}} which is a wrapper for \code{\link[officer:read_docx]{officer::read_docx()}},
\code{\link[officer:read_pptx]{officer::read_pptx()}}, and \code{\link[officer:read_xlsx]{officer::read_xlsx()}}. This function uses both a
filename and path (the original officer functions only use a path) and
default to use a rdocx, rpptx, or rxlsx class object if provided.
}
\seealso{
\code{\link[officer:read_docx]{officer::read_docx()}}
}
