% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_officer.R
\name{reduce_officer}
\alias{reduce_officer}
\title{Reduce a list to a single officer object}
\usage{
reduce_officer(
  x = NULL,
  .f = function(x, value, ...) {
     vec_add_to_body(x, value = value, ...)
 },
  value = NULL,
  ...,
  .write_path = NULL
)
}
\arguments{
\item{x}{File path or officer object.}

\item{.f}{Any function taking an officer object as the first parameter and a
value as the second parameter, Defaults to anonymous function:
\code{function(x, value, ...) { vec_add_to_body(x, value = value, ...) }}}

\item{value}{A vector of values that are support by the function passed to
.f, Default: \code{NULL}}

\item{...}{Additional parameters passed to \code{\link[purrr:reduce]{purrr::reduce()}}.}

\item{.write_path}{If .write_path not \code{NULL}, it should be a file path that
is passed to \code{\link[=write_officer]{write_officer()}}, allowing you to modify a docx file and
write it back to a file in a single function call.}
}
\value{
A rdocx, rpptx, or rxlsx object.
}
\description{
\code{\link[=reduce_officer]{reduce_officer()}} is a wrapper for \code{\link[purrr:reduce]{purrr::reduce()}}.
}
\examples{
\dontrun{
if(interactive()){
 x <- reduce_officer(value = LETTERS)

 officer_summary(x)
 }
}
}
\seealso{
\code{\link[purrr:reduce]{purrr::reduce()}}, \code{\link[=vec_add_to_body]{vec_add_to_body()}}
}
