% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_docx.R
\name{save_rdocx}
\alias{save_rdocx}
\title{Save rdocx object to docx file}
\usage{
save_rdocx(docx, path, overwrite = TRUE)
}
\arguments{
\item{docx}{A rdocx object to save.}

\item{path}{File path.}

\item{overwrite}{If \code{FALSE} and file exists, abort save function. If \code{TRUE},
save file even if file exists at path.}
}
\description{
Save rdocx object to docx file
}
