% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_officer.R
\name{set_office_path}
\alias{set_office_path}
\title{Set filepath for docx file}
\usage{
set_office_path(
  filename = NULL,
  path = NULL,
  fileext = c("docx", "pptx", "xlsx")
)
}
\description{
Set filepath for docx file
}
\keyword{internal}
