% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{write_docx}
\alias{write_docx}
\title{Write a rdocx object to a docx file}
\usage{
write_docx(docx, path, overwrite = TRUE)
}
\arguments{
\item{docx}{A rdocx object to save.}

\item{path}{File path.}

\item{overwrite}{If \code{FALSE} and file exists, abort save function. If \code{TRUE},
save file even if file exists at path.}
}
\description{
Write a rdocx object to a docx file
}
