% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_body.R
\name{add_to_body}
\alias{add_to_body}
\alias{add_text_to_body}
\alias{add_xml_to_body}
\alias{add_gt_to_body}
\title{Add an xml string or paragraph of text at a specified position in a Word document}
\usage{
add_to_body(
  docx,
  keyword = NULL,
  id = NULL,
  index = NULL,
  value = NULL,
  str = NULL,
  ...
)

add_text_to_body(
  docx,
  value,
  style = NULL,
  .na = "NA",
  .null = NULL,
  .envir = parent.frame(),
  ...
)

add_xml_to_body(docx, str = NULL, pos = "after", ...)

add_gt_to_body(docx, gt_object, pos = "after", ...)
}
\arguments{
\item{docx}{A rdocx object.}

\item{keyword, id}{A keyword string used to place cursor with
\code{\link[officer:cursor]{officer::cursor_reach()}} or bookmark id with \code{\link[officer:cursor]{officer::cursor_bookmark()}}.
Defaults to \code{NULL}. If keyword or id are not provided, the gt object is
inserted at the front of the document.}

\item{index}{A integer matching a doc_index value appearing in a summary of
the docx object created with \code{\link[officer:docx_summary]{officer::docx_summary()}}. If index is for a
paragraph value, the text of the pargraph is used as a keyword.}

\item{value}{object to add in the document. Supported objects
are vectors, data.frame, graphics, block of formatted paragraphs,
unordered list of formatted paragraphs,
pretty tables with package flextable, 'Microsoft' charts with package mschart.}

\item{str}{a wml string}

\item{...}{Additional parameters passed to \code{\link[officer:body_add]{officer::body_add()}} or
\code{\link[officer:body_add_xml]{officer::body_add_xml()}}.}

\item{style}{paragraph style name. These names are available with function \link[officer]{styles_info}
and are the names of the Word styles defined in the base document (see
argument \code{path} from \link[officer]{read_docx}).}

\item{.na}{[\code{character(1)}: \sQuote{NA}]\cr Value to replace \code{NA} values
with. If \code{NULL} missing values are propagated, that is an \code{NA} result will
cause \code{NA} output. Otherwise the value is replaced by the value of \code{.na}.}

\item{.null}{[\code{character(1)}: \sQuote{character()}]\cr Value to replace
NULL values with. If \code{character()} whole output is \code{character()}. If
\code{NULL} all NULL values are dropped (as in \code{paste0()}). Otherwise the
value is replaced by the value of \code{.null}.}

\item{.envir}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored. If \code{NULL} is passed, it is equivalent to \code{\link[=emptyenv]{emptyenv()}}.}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{gt_object}{A gt object converted to an OOXML string with
\code{\link[gt:as_word]{gt::as_word()}} then passed to \code{\link[=add_xml_to_body]{add_xml_to_body()}}. Required.}
}
\value{
A rdocx object with added xml, gt tables, or paragraphs of text.
}
\description{
Wrappers for \code{\link[officer:body_add]{officer::body_add()}} and \code{\link[officer:body_add_xml]{officer::body_add_xml()}} that use the
\code{\link[=cursor_docx]{cursor_docx()}} helper functions to allow input of filename and path and use
of multiple options for cursor placement. \code{\link[=add_text_to_body]{add_text_to_body()}} passed value
to \code{\link[glue:glue]{glue::glue()}} to add support for glue string interpolation.
\code{\link[=add_gt_to_body]{add_gt_to_body()}} converts gt tables to OOXML with \code{\link[gt:as_word]{gt::as_word()}}.
}
