% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{check_office_fileext}
\alias{check_office_fileext}
\alias{check_docx_fileext}
\alias{check_pptx_fileext}
\alias{check_xlsx_fileext}
\title{Check file path for a docx, pptx, or xlsx file extension}
\usage{
check_office_fileext(
  x,
  arg = caller_arg(x),
  fileext = c("docx", "pptx", "xlsx"),
  call = parent.frame(),
  ...
)

check_docx_fileext(x, arg = caller_arg(x), ..., .envir = current_env())

check_pptx_fileext(x, arg = caller_arg(x), ..., .envir = current_env())

check_xlsx_fileext(x, arg = caller_arg(x), ..., .envir = current_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name of object to check. Used to improve
\code{\link[cli:cli_abort]{cli::cli_abort()}} messages. Defaults to \code{caller_arg(x)}.}

\item{fileext}{File extensions to allow without error. Defaults to "docx",
"pptx", "xlsx".}

\item{call}{The execution environment of a currently running function passed
to \code{\link[cli:cli_abort]{cli::cli_abort()}}. Defaults to \code{parent.frame()}}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\description{
Check file path for a docx, pptx, or xlsx file extension
}
