% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{check_officer}
\alias{check_officer}
\alias{check_docx}
\alias{check_pptx}
\alias{check_xlsx}
\title{Check if x is a rdocx, rpptx, or rxlsx object}
\usage{
check_officer(x, arg = caller_arg(x), what = c("rdocx", "rpptx", "rxlsx"), ...)

check_docx(x, arg = caller_arg(x), ...)

check_pptx(x, arg = caller_arg(x), ...)

check_xlsx(x, arg = caller_arg(x), ...)
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name for x. Defaults to \code{caller_arg(x)}.}

\item{what}{Class names to check}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\description{
Check if x is a rdocx, rpptx, or rxlsx object
}
