% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{check_officer}
\alias{check_officer}
\alias{check_docx}
\alias{check_pptx}
\alias{check_xlsx}
\alias{check_block_list}
\title{Check if x is a rdocx, rpptx, or rxlsx object}
\usage{
check_officer(
  x,
  arg = caller_arg(x),
  what = c("rdocx", "rpptx", "rxlsx"),
  call = caller_env(),
  ...
)

check_docx(x, arg = caller_arg(x), call = caller_env(), ...)

check_pptx(x, arg = caller_arg(x), call = caller_env(), ...)

check_xlsx(x, arg = caller_arg(x), call = caller_env(), ...)

check_block_list(
  x,
  arg = caller_arg(x),
  allow_empty = FALSE,
  allow_null = FALSE,
  call = caller_env()
)
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name of object to check. Used to improve
\code{\link[cli:cli_abort]{cli::cli_abort()}} messages. Defaults to \code{caller_arg(x)}.}

\item{what}{Class names to check}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}

\item{allow_empty}{If \code{TRUE}, \code{\link[=check_block_list]{check_block_list()}} allows an empty block
list.}
}
\description{
Check if x is a rdocx, rpptx, or rxlsx object
}
