% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_summary.R
\name{check_officer_summary}
\alias{check_officer_summary}
\title{Check a summary data.frame created from a rdocx or rpptx object}
\usage{
check_officer_summary(
  x,
  n = NULL,
  content_type = NULL,
  summary_type = "doc",
  tables = FALSE,
  ...,
  arg = caller_arg(x),
  call = parent.frame()
)
}
\arguments{
\item{x}{An object to check if it is a data.frame object created with
\code{\link[officer:docx_summary]{officer::docx_summary()}} or another summary function.}

\item{n}{Required number of rows. Optional. If n is more than length 1,
checks to make sure the number of rows is within the range of \code{max(n) }and
\code{min(n)}. Defaults to \code{NULL}.}

\item{content_type}{Required content_type, e.g. "paragraph", "table cell", or
"image". Optional. Defaults to \code{NULL}.}

\item{summary_type}{Summary type. Options "doc", "docx", "pptx", "slide", or
"layout". "doc" requires the data.frame include a "content_type" column but
allows columns for either a docx or pptx summary.}

\item{tables}{If \code{TRUE}, require that the summary include the column names
indicated a table is present in the rdocx or rpptx summary.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}

\item{arg}{Argument name to use in error messages. Defaults to
\code{caller_arg(x)}}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\value{
Invisibly returns \code{NULL} if \code{x} passes the checks. Otherwise,
errors.
}
\description{
Check an object and error if an object is not a data.frame with the required
column names to be a summary data.frame created from a rdocx or rpptx object.
Optionally check for the number of rows, a specific content_type value, or if
tables are included in the document the summary was created from.
}
\seealso{
Other summary functions:
\code{\link[=officer_summary]{officer_summary()}},
\code{\link[=officer_summary_levels]{officer_summary_levels()}}
}
\concept{summary functions}
