% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursor_docx.R
\name{cursor_docx}
\alias{cursor_docx}
\title{Set cursor position in rdocx object based on keyword, id, or index}
\usage{
cursor_docx(docx, keyword = NULL, id = NULL, index = NULL)
}
\arguments{
\item{docx}{A rdocx object.}

\item{keyword, id}{A keyword string used to place cursor with
\code{\link[officer:cursor]{officer::cursor_reach()}} or bookmark id with \code{\link[officer:cursor]{officer::cursor_bookmark()}}.
Defaults to \code{NULL}. If keyword or id are not provided, the gt object is
inserted at the front of the document.}

\item{index}{A integer matching a doc_index value appearing in a summary of
the docx object created with \code{\link[officer:docx_summary]{officer::docx_summary()}}. If index is for a
paragraph value, the text of the pargraph is used as a keyword.}
}
\description{
Set cursor position in rdocx object based on keyword, id, or index
}
\seealso{
\code{\link[officer:cursor]{officer::cursor_begin()}}, \code{\link[officer:docx_summary]{officer::docx_summary()}}
}
