% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims_docx.R
\name{dims_docx_ext}
\alias{dims_docx_ext}
\title{Get page and block layout dimensions from an rdocx object}
\usage{
dims_docx_ext(docx)
}
\arguments{
\item{docx}{A rdocx object to get dimensions for.}
}
\description{
This function extends \code{\link[officer:docx_dim]{officer::docx_dim()}} by also returning the body text
dimensions within the margins, the aspect ratio of the page and body, and the
page orientation as a string ("landscape" or "portrait").
}
\seealso{
\code{\link[officer:docx_dim]{officer::docx_dim()}}
}
