% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docx.R
\name{has_docx_fileext}
\alias{has_docx_fileext}
\title{Does x have a docx file extension?}
\usage{
has_docx_fileext(x = NULL)
}
\arguments{
\item{x}{A character vector to check for matches, or an object which can be
coerced by \code{\link[=as.character]{as.character()}} to a character vector.}
}
\description{
Does x have a docx file extension?
}
\keyword{internal}
