% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_officer.R
\name{is_officer}
\alias{is_officer}
\alias{is_rdocx}
\alias{is_rpptx}
\alias{is_block_list}
\title{Is this object an officer class object?}
\usage{
is_officer(x, what = c("rdocx", "rpptx", "rxlsx"))

is_rdocx(x)

is_rpptx(x)

is_block_list(x)
}
\description{
\code{\link[=is_officer]{is_officer()}} and variants are basic wrappers for \code{\link[=inherits]{inherits()}} to check for
object classes created by functions from the officer package. These functions
are intended for internal use but are exposed for use by other extension
developers.
}
