% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_media.R
\name{officer_media}
\alias{officer_media}
\title{Copy media from a docx or pptx file to a folder}
\usage{
officer_media(filename = NULL, path = NULL, dest = "media", overwrite = TRUE)
}
\arguments{
\item{dest}{Folder name or path to copy media files to. dest is created if no
folder exists at that location.}

\item{overwrite}{If \code{TRUE} (default), overwrite any files with the same names
at the dest location.}
}
\description{
Copy media from a docx or pptx file to a folder
}
