% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer-utils.R
\name{officer_open}
\alias{officer_open}
\title{Preview a rdocx, rpptx, or rxlsx object in local default applications}
\usage{
officer_open(
  x,
  path = NULL,
  ...,
  overwrite = FALSE,
  interactive = is_interactive()
)
}
\arguments{
\item{x}{A rdocx, rpptx, or rxlsx object to save. Document properties won't
be set for rxlsx objects.}

\item{path}{Optional. Set as temporary file with file extension matching type of input object x.}

\item{overwrite}{Defaults to \code{FALSE}.}
}
\value{
Input object x
}
\description{
Preview a rdocx, rpptx, or rxlsx object in local default applications
}
