% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_officer.R
\name{officer_properties}
\alias{officer_properties}
\title{Get doc properties for a rdocx or rpptx object}
\usage{
officer_properties(x, values = list(), keep.null = FALSE)
}
\description{
Get doc properties for a rdocx or rpptx object
}
\keyword{internal}
