% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_summary.R
\name{officer_summary}
\alias{officer_summary}
\title{Summarize a rdocx or rpptx object}
\usage{
officer_summary(x)
}
\arguments{
\item{x}{A rdocx or rpptx object passed to \code{\link[officer:docx_summary]{officer::docx_summary()}} or \code{\link[officer:pptx_summary]{officer::pptx_summary()}}.}
}
\value{
A data.frame object.
}
\description{
Summarize a rdocx or rpptx object
}
