% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_levels.R
\name{officer_summary_levels}
\alias{officer_summary_levels}
\title{Create new columns to an officer summary data.frame based on specific levels}
\usage{
officer_summary_levels(
  x,
  levels = NULL,
  levels_from = "style_name",
  exclude_levels = NULL,
  fill_col = "text",
  direction = c("down", "up", "downup", "updown"),
  ...,
  call = caller_env()
)
}
\arguments{
\item{x}{A input data.frame (assumed to be from \code{\link[=officer_summary]{officer_summary()}} for
default values).}

\item{levels}{Levels to use. If \code{NULL} (default), levels is set to unique,
non-NA values from the levels_from column.}

\item{levels_from}{Column name to use for identifying levels. Defaults to
"style_name".}

\item{exclude_levels}{Levels to exclude from process of adding new columns.}

\item{fill_col}{Name of column to fill , Defaults to "text".}

\item{direction}{Direction of fill passed to \code{\link[vctrs:vec_fill_missing]{vctrs::vec_fill_missing()}},
Default: c("down", "up", "downup", "updown")}

\item{...}{
  Arguments passed on to \code{\link[=officer_summary]{officer_summary}}
  \describe{
    \item{\code{summary_type}}{Summary type. Options "doc", "docx", "pptx", "slide", or
"layout". "doc" requires the data.frame include a "content_type" column but
allows columns for either a docx or pptx summary.}
    \item{\code{index}}{slide index}
  }}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
This function works with \code{\link[=fill_with_pattern]{fill_with_pattern()}} to help convert hierarchically
organized text, e.g. text with leveled headings or other styles into a
data.frame where new columns hold the value of the preceding (or succeeding)
heading text.
}
