% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_tables.R
\name{officer_tables}
\alias{officer_tables}
\title{Get tables from a rdocx or rpptx object}
\usage{
officer_tables(x, index = NULL, has_header = TRUE)
}
\arguments{
\item{x}{A rdocx or rpptx object or a data.frame created with
\code{\link[=officer_summary]{officer_summary()}}}

\item{index}{A index value matching a doc_index value for a table in the
summary data.frame, Default: \code{NULL}}

\item{has_header}{If \code{TRUE}, tables are expected to have implicit headers
even if the object Default: \code{TRUE}}
}
\value{
A data.frame or list of data.frame (or list objects).
}
\description{
Get one or more tables from a rdocx or rpptx object. Functions based on
example code on extracting Word document and PowerPoint slides in the
\href{https://ardata-fr.github.io/officeverse/extract-content.html#word-tables}{officeverse documentation}.
}
\examples{
docx <- read_docx_ext(
  filename = "example.docx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(docx)

pptx <- read_pptx_ext(
  filename = "example.pptx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(pptx)[[1]]

}
