% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerExtras-package.R
\docType{package}
\name{officerExtras-package}
\alias{officerExtras}
\alias{officerExtras-package}
\title{officerExtras: Extra Helpers for 'officer'}
\description{
Helper and convenience functions using the 'officer' package to modify docx files.
}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
