% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docs_url.R
\name{read_docs_url}
\alias{read_docs_url}
\title{Read a Google Docs document, Slides, or Sheets URL to a rdocx, rpptx, or
rxlsx object}
\usage{
read_docs_url(url, format = NULL, filename = NULL, path = NULL, quiet = TRUE)
}
\arguments{
\item{url}{A URL for a Google Doc, Google Slides presentation, or Google
Sheets.}

\item{format}{File format to use for export URL (typically set
automatically). Options are \code{NULL} (default), "doc", "pptx", or "xslx".
"pdf" and "csv" be supported in the future.}

\item{filename}{Destination file name. Optional. If filename is \code{NULL},
downloaded file is removed as part of the function execution.}

\item{path}{Folder path. Optional.}

\item{quiet}{If \code{TRUE}, suppress messages when downloading file.}
}
\description{
Uses the Google Docs, Slides, or Sheets URL to export a file locally, read to
an officer object. If filename is \code{NULL}, exported file is removed after
export.
}
