% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_add_to_body.R
\name{vec_add_to_body}
\alias{vec_add_to_body}
\title{Add a vector of objects to a rdocx object using \code{add_to_body()}}
\usage{
vec_add_to_body(docx, ..., .size = NULL, .call = caller_env())
}
\arguments{
\item{docx}{A rdocx object.}

\item{...}{
  Arguments passed on to \code{\link[=add_to_body]{add_to_body}}
  \describe{
    \item{\code{keyword,id}}{A keyword string used to place cursor with
\code{\link[officer:cursor]{officer::cursor_reach()}} or bookmark id with \code{\link[officer:cursor]{officer::cursor_bookmark()}}.
Defaults to \code{NULL}. If keyword or id are not provided, the gt object is
inserted at the front of the document.}
    \item{\code{index}}{A integer matching a doc_index value appearing in a summary of
the docx object created with \code{\link[officer:docx_summary]{officer::docx_summary()}}. If index is for a
paragraph value, the text of the pargraph is used as a keyword.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
    \item{\code{value}}{object to add in the document. Supported objects
are vectors, data.frame, graphics, block of formatted paragraphs,
unordered list of formatted paragraphs,
pretty tables with package flextable, 'Microsoft' charts with package mschart.}
    \item{\code{style}}{paragraph style name. These names are available with function \link[officer]{styles_info}
and are the names of the Word styles defined in the base document (see
argument \code{path} from \link[officer]{read_docx}).}
    \item{\code{str}}{a wml string}
    \item{\code{pos}}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
  }}

\item{.size}{Desired output size.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=vec_add_to_body]{vec_add_to_body()}} is a vectorized variant \code{\link[=add_to_body]{add_to_body()}} that allows users
to supply a vector of values or str inputs to add multiple blocks of text,
images, plots, or tables to a document. Alternatively, the function also
supports adding a single object at multiple locations or using multiple
different styles.  All parameters are recycled using
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}} so inputs must be length 1 or match the length
of the longest input vector.
}
\examples{
docx_example <- read_officer()

docx_example <- vec_add_to_body(
  docx_example,
  value = "Sample text",
  style = c("heading 1", "heading 2", "Normal")
)

officer_summary(docx_example)
}
