% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_officer.R
\name{write_officer}
\alias{write_officer}
\title{Write a rdocx, rpptx, or rxlsx object to a file}
\usage{
write_officer(x, path, overwrite = TRUE, modified_by = Sys.getenv("USER"), ...)
}
\arguments{
\item{x}{A rdocx, rpptx, or rxlsx object to save. Document properties won't
be set for rxlsx objects.}

\item{path}{File path.}

\item{overwrite}{If \code{TRUE} (default), remove file at path if it already
exists. If \code{FALSE} and file exists, this function aborts.}

\item{modified_by}{If the withr package is installed, modified_by overrides
the default value for the lastModifiedBy property assigned to the output
file by officer. Defaults to \code{Sys.getenv("USER")} (the same value used by
officer).}

\item{...}{
  Arguments passed on to \code{\link[officer:set_doc_properties]{officer::set_doc_properties}}
  \describe{
    \item{\code{title,subject,creator,description}}{text fields}
    \item{\code{created}}{a date object}
    \item{\code{values}}{a named list (names are field names), each element is a single
character value specifying value associated with the corresponding field name.
If \code{values} is provided, argument \code{...} will be ignored.}
  }}
}
\value{
Returns the input object (invisibly) and writes the rdocx, rpptx, or
rxlsx object to a file with a name and location matching the provided path.
}
\description{
Write a rdocx, rpptx, or rxlsx object to a file
}
